requirements.txt
langchain==0.0.208
deeplake==3.8.11
openai==0.27.8
tiktoken==0.4.0
selenium==4.15.2
huggingface
newspaper3k
python-dotenv
streamlit
transformers
cohere
google-api-python-client
googlesearch-python
newspaper3k
google-generativeai
unstructured
SeleniumURLLoader
pypdf
chromadb
pyvis
chromium
libmagic
python-magic
python-magic-bin
sklearn
scipy
numpy
sentence-transformer
apify-client
yt_dlp
openai-whisper
#gradio